# openapi.yaml
openapi: 3.1.0
info:
  title: Keyforge API
  version: 0.1.0
  description: |
    Multi-tenant VaultWarden management API for Kubernetes.
    
    Keyforge provisions isolated VaultWarden instances in Kubernetes and manages
    organizations with shared password vaults for teams.
  contact:
    name: Keyforge
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /instances/{instance_id}/organisations:
    post:
      summary: Create a new organisation
      description: |
        Creates a new organisation within an instance. This:
        1. Registers a service account user in VaultWarden
        2. Authenticates the user
        3. Creates a VaultWarden organization
        Organizations enable multi-user password sharing.
      tags:
        - organisations
      parameters:
        - $ref: '#/components/parameters/InstanceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganisationRequest'
      responses:
        '201':
          description: Organisation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrganisationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Instance not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /instances/{instance_id}/organisations/{organisation_id}:
    get:
      summary: Get organisation details
      description: Returns details for a specific organisation
      tags:
        - organisations
      parameters:
        - $ref: '#/components/parameters/InstanceId'
        - $ref: '#/components/parameters/OrganisationId'
      responses:
        '200':
          description: Organisation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organisation'
        '404':
          description: Instance or organisation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /instances/{instance_id}/organisations/{organisation_id}/passwords:
    post:
      summary: Create a new password
      description: |
        Creates a new password entry (cipher) in the organisation's VaultWarden organization.
        The password is stored encrypted and accessible to all organization members.
      tags:
        - organisations
      parameters:
        - $ref: '#/components/parameters/InstanceId'
        - $ref: '#/components/parameters/OrganisationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePasswordRequest'
      responses:
        '201':
          description: Password created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePasswordResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Instance or organisation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Organisation not properly initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      summary: List passwords
      description: Returns all passwords for an organisation
      tags:
        - organisations
      parameters:
        - $ref: '#/components/parameters/InstanceId'
        - $ref: '#/components/parameters/OrganisationId'
      responses:
        '200':
          description: List of passwords
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordListResponse'
        '404':
          description: Instance or organisation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Organisation not properly initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /instances/{instance_id}/organisations/{organisation_id}/passwords/{password_id}:
    get:
      summary: Get password details
      description: Returns the full password details including the decrypted password value
      tags:
        - organisations
      parameters:
        - $ref: '#/components/parameters/InstanceId'
        - $ref: '#/components/parameters/OrganisationId'
        - $ref: '#/components/parameters/PasswordId'
      responses:
        '200':
          description: Password details with value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordDetail'
        '404':
          description: Instance, organisation, or password not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Organisation not properly initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Update a password
      description: Updates the name and/or value of a password
      tags:
        - organisations
      parameters:
        - $ref: '#/components/parameters/InstanceId'
        - $ref: '#/components/parameters/OrganisationId'
        - $ref: '#/components/parameters/PasswordId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordDetail'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Instance, organisation, or password not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Organisation not properly initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Delete a password
      description: Deletes a password entry from the organisation
      tags:
        - organisations
      parameters:
        - $ref: '#/components/parameters/InstanceId'
        - $ref: '#/components/parameters/OrganisationId'
        - $ref: '#/components/parameters/PasswordId'
      responses:
        '200':
          description: Password deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password deleted successfully
        '404':
          description: Instance, organisation, or password not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Organisation not properly initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    InstanceId:
      name: instance_id
      in: path
      required: true
      description: Unique identifier for the instance
      schema:
        type: string
        pattern: '^union-[a-f0-9]{16}$'
        example: union-2574af3733dd26f5
    OrganisationId:
      name: organisation_id
      in: path
      required: true
      description: Unique identifier for the organisation
      schema:
        type: string
        pattern: '^society-[a-f0-9]{16}$'
        example: society-e6557cc8e1656983
    PasswordId:
      name: password_id
      in: path
      required: true
      description: Unique identifier for the password
      schema:
        type: string
        pattern: '^pwd-[a-f0-9]{16}$'
        example: pwd-657bbc9ee11296d0
  schemas:
...
    CreateOrganisationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          description: Human-readable name for the organisation
          example: engineering-team
    CreateOrganisationResponse:
      type: object
      properties:
        organisation_id:
          type: string
          example: society-e6557cc8e1656983
        instance_id:
          type: string
          example: union-2574af3733dd26f5
        vaultwd_org_id:
          type: string
          format: uuid
          description: VaultWarden organization UUID
          example: 42b2b981-ed49-48ce-9b3f-7ad6b53c2e46
        status:
          $ref: '#/components/schemas/OrganisationStatus'
    Organisation:
      type: object
      properties:
        id:
          type: string
          example: society-e6557cc8e1656983
        name:
          type: string
          example: engineering-team
        instance_id:
          type: string
          example: union-2574af3733dd26f5
        vaultwd_org_id:
          type: string
          format: uuid
          example: 42b2b981-ed49-48ce-9b3f-7ad6b53c2e46
        vaultwd_user_email:
          type: string
          format: email
          description: Service account email in VaultWarden
          example: society-e6557cc8e1656983@keyforge.local
        vaultwd_user_token:
          type: string
          description: Bearer token for VaultWarden API calls
        status:
          $ref: '#/components/schemas/OrganisationStatus'
        created_at:
          type: integer
          format: int64
          example: 1761318156000
    OrganisationStatus:
      type: string
      enum:
        - pending
        - created
        - failed
      description: |
        - pending: Organisation creation in progress
        - created: Organization created successfully
        - failed: Creation failed
    CreatePasswordRequest:
      type: object
      required:
        - name
        - value
      properties:
        name:
          type: string
          minLength: 1
          description: Name/label for the password entry
          example: database-password
        value:
          type: string
          minLength: 1
          description: The actual password value (stored encrypted in VaultWarden)
          example: super-secret-123
    CreatePasswordResponse:
      type: object
      properties:
        password_id:
          type: string
          example: pwd-657bbc9ee11296d0
        organisation_id:
          type: string
          example: society-e6557cc8e1656983
        name:
          type: string
          example: database-password
        created_at:
          type: integer
          format: int64
          example: 1761318165804
    Password:
      type: object
      properties:
        id:
          type: string
          example: pwd-657bbc9ee11296d0
        organisation_id:
          type: string
          example: society-e6557cc8e1656983
        name:
          type: string
          example: database-password
        created_at:
          type: integer
          format: int64
          example: 1761318165804
    PasswordDetail:
      type: object
      properties:
        id:
          type: string
          example: pwd-657bbc9ee11296d0
        organisation_id:
          type: string
          example: society-e6557cc8e1656983
        name:
          type: string
          example: database-password
        value:
          type: string
          description: The decrypted password value
          example: super-secret-123
        created_at:
          type: integer
          format: int64
          example: 1761318165804
    UpdatePasswordRequest:
      type: object
      properties:
        name:
          type: string
          description: New name for the password (optional)
          example: updated-password-name
        value:
          type: string
          description: New password value (optional)
          example: new-super-secret-456
      minProperties: 1
    PasswordListResponse:
      type: object
      properties:
        organisation_id:
          type: string
          example: society-e6557cc8e1656983
        passwords:
          type: array
          items:
            $ref: '#/components/schemas/Password'
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Union not found